+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.13.1",
     "react-chartjs-2": "^2.10.0",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
     "serialize-javascript": "^4.0.0",
     "tailwindcss": "^1.8.4",
